pyproject.toml
[project]
name = "intelligent-claims-orchestrator"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "aws-cdk-lib>=2.235.1,<3.0.0",
    "boto3>=1.42.30",
    "constructs>=10.0.0,<11.0.0",
    "pillow>=12.1.0",
    "pip>=25.3",
    "reportlab>=4.4.9",
    "requests>=2.32.5",
]

[dependency-groups]
dev = [
    "pytest==8.4.2",
]
